@@ -1,16 +1,12 @@
 import React from 'react';
 import { useAuth0 } from '@auth0/auth0-react';
-import { createBrowserHistory } from 'history';
-import { Route, Router, Switch } from 'react-router-dom';
+import { Route, Routes } from 'react-router-dom';
 import './App.css';
-import { ProtectedRoute } from './ProtectedRoute';
 import { Nav } from './Nav';
 import { Error } from './Error';
 import { Loading } from './Loading';
 import { Users } from './Users';
-
-// Use `createHashHistory` to use hash routing
-export const history = createBrowserHistory();
+import { ProtectedRoute } from './ProtectedRoute';
 
 function App() {
   const { isLoading, error } = useAuth0();
@@ -20,14 +16,14 @@ function App() {
   }
 
   return (
-    <Router history={history}>
+    <>
       <Nav />
       {error && <Error message={error.message} />}
-      <Switch>
-        <Route path="/" exact />
-        <ProtectedRoute path="/users" component={Users} />
-      </Switch>
-    </Router>
+      <Routes>
+        <Route path="/" />
+        <Route path="/users" element={<ProtectedRoute component={Users} />} />
+      </Routes>
+    </>
   );
 }
 
