@@ -1,17 +1,12 @@
-import React, { useEffect, useState } from 'react';
-import { useHistory, Link } from 'react-router-dom';
+import React from 'react';
+import { useLocation, Link } from 'react-router-dom';
 import { useAuth0 } from '@auth0/auth0-react';
 
 export function Nav() {
   const { isAuthenticated, user, loginWithRedirect, logout } = useAuth0<{
     name: string;
   }>();
-  const history = useHistory();
-  const [pathname, setPathname] = useState(() => history.location.pathname);
-
-  useEffect(() => {
-    return history.listen(({ pathname }) => setPathname(pathname));
-  }, [history]);
+  const { pathname } = useLocation();
 
   return (
     <nav className="navbar navbar-expand-lg navbar-light bg-light">
