@@ -1,27 +1,40 @@
-import React from 'react';
+import React, { PropsWithChildren } from 'react';
 import ReactDOM from 'react-dom';
-import App, { history } from './App';
+import App from './App';
 import { Auth0Provider, AppState } from '@auth0/auth0-react';
+import { BrowserRouter, useNavigate } from 'react-router-dom';
+import { Auth0ProviderOptions } from '../../../src';
 
-const onRedirectCallback = (appState?: AppState) => {
-  // If using a Hash Router, you need to use window.history.replaceState to
-  // remove the `code` and `state` query parameters from the callback url.
-  // window.history.replaceState({}, document.title, window.location.pathname);
-  history.replace((appState && appState.returnTo) || window.location.pathname);
+const Auth0ProviderWithRedirectCallback = ({
+  children,
+  ...props
+}: PropsWithChildren<Auth0ProviderOptions>) => {
+  const navigate = useNavigate();
+
+  const onRedirectCallback = (appState?: AppState) => {
+    navigate((appState && appState.returnTo) || window.location.pathname);
+  };
+
+  return (
+    <Auth0Provider onRedirectCallback={onRedirectCallback} {...props}>
+      {children}
+    </Auth0Provider>
+  );
 };
 
 ReactDOM.render(
   <React.StrictMode>
-    <Auth0Provider
-      domain={process.env.REACT_APP_DOMAIN}
-      clientId={process.env.REACT_APP_CLIENT_ID}
-      redirectUri={window.location.origin}
-      audience={process.env.REACT_APP_AUDIENCE}
-      scope="read:users"
-      onRedirectCallback={onRedirectCallback}
-    >
-      <App />
-    </Auth0Provider>
+    <BrowserRouter>
+      <Auth0ProviderWithRedirectCallback
+        domain={process.env.REACT_APP_DOMAIN}
+        clientId={process.env.REACT_APP_CLIENT_ID}
+        redirectUri={window.location.origin}
+        audience={process.env.REACT_APP_AUDIENCE}
+        scope="read:users"
+      >
+        <App />
+      </Auth0ProviderWithRedirectCallback>
+    </BrowserRouter>
   </React.StrictMode>,
   document.getElementById('root')
 );
